Move the code generation to their own crate
